Change color depth to 8 bits to free SRAM 0.2.2->0.2.3
With 0.2.2 the SRAM stack heap minimum reaches 2.748kB when Wi Fi is operational. This low RAM makes the pairing and operation of the Wi/Fi features to fail or be slow. Change: to release RAM, the sprite for graphics is moved from 16bits depth to 8bits, effectively releasing 57.6kB of RAM. Unfortunately, using 8bit sprite does not allow the use of the DMA feature, those needed to make a regression to use again sprite.push. This increases the time to render one screen back to 40ms.
On 0.2.3 the minimum stack depth is 53.528kB and the behavior of Wi/Fi return to normalcy.
- Other minor changes.